+++ app/crud/booking.py
@@ -1,34 +1,11 @@
 from sqlalchemy.orm import Session
 from fastapi import HTTPException
-from . import models, schemas
-
-# ユーザー一覧取得
-def get_users(db: Session, skip: int = 0, limit: int = 100):
-    return db.query(models.User).offset(skip).limit(limit).all()
-
-# 会議室一覧取得
-def get_rooms(db: Session, skip: int = 0, limit: int = 100):
-    return db.query(models.Room).offset(skip).limit(limit).all()
+from app.models import models
+from app.schemas import schemas
 
 # 予約一覧取得
 def get_bookings(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.Booking).offset(skip).limit(limit).all()
-
-# ユーザー登録
-def create_user(db: Session, user: schemas.User):
-    db_user = models.User(username=user.username)
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
-    return db_user
-
-# 会議室登録
-def create_room(db: Session, room: schemas.Room):
-    db_room = models.Room(room_name=room.room_name, capacity=room.capacity)
-    db.add(db_room)
-    db.commit()
-    db.refresh(db_room)
-    return db_room
 
 # 予約登録
 def create_booking(db: Session, booking: schemas.Booking):
