+++ app/crud/room.py
@@ -1,0 +1,15 @@
+from sqlalchemy.orm import Session
+from app.models import models
+from app.schemas import schemas
+
+# 会議室一覧取得
+def get_rooms(db: Session, skip: int = 0, limit: int = 100):
+    return db.query(models.Room).offset(skip).limit(limit).all()
+
+# 会議室登録
+def create_room(db: Session, room: schemas.Room):
+    db_room = models.Room(room_name=room.room_name, capacity=room.capacity)
+    db.add(db_room)
+    db.commit()
+    db.refresh(db_room)
+    return db_room
