+++ app/crud/user.py
@@ -1,0 +1,15 @@
+from sqlalchemy.orm import Session
+from app.models import models
+from app.schemas import schemas
+
+# ユーザー一覧取得
+def get_users(db: Session, skip: int = 0, limit: int = 100):
+    return db.query(models.User).offset(skip).limit(limit).all()
+
+# ユーザー登録
+def create_user(db: Session, user: schemas.User):
+    db_user = models.User(username=user.username)
+    db.add(db_user)
+    db.commit()
+    db.refresh(db_user)
+    return db_user
